requirements.txt
urwid
requests
python-dotenv
pillow
numpy
